Introduce a testing infrastructure for KUNIT: A Kernel Unit Testing Framework.
This commit introduces the kunit directory and the runtest.sh that is needed to execute these tests.

This method of testing KUNIT does not rely on UML like the attempt being produced by the KernelCI project

To get all these tests to run, the kernel must have CONFIG_KUNIT=m and CONFIG_KUNIT_ALL_TESTS=m.

New tests must be added to the list at the top of the runtests.sh. The test names are the names of the kunit modules.