Optionally install clang dependencies
We're trying to test some kernels built with clang. That means during
the testing, we also need to build any external modules with clang.

We still have to figure out how to pass along the information about what
compiler should be used for the rebuilds, but let's start with actually
getting the dependencies in place.

Signed-off-by: Veronika Kabatova <vkabatov@redhat.com>