perf: multiple changes in internal-testsuite/runtest.sh
(1): It seems that when ARCH variable is set, it disturbs kbuild search
process in the perf LLVM tests, so ARCH variable has to be undefined in
order not to disturb the script.

(2): Install python3-perf to make sure that some python*-perf is
installed on the box.

(3): On systems with more CPUs, the defaul "ulimit -l" value is not
enough.

(4): Install more packages to run the BPF tests properly.

(5): For easier debugging run the tests in no-fork mode.

(6): The parsing with the RESULT variable did not contain results of
test subtests.