Make modifications to reduce the execution time of the job
The san-device-stress job is occasionally hitting the local watchdog
as the test execution is sometimes taking over 3 hours. To decrease the
time needed to execute this test, we will change the size of the FIO jobs
to 1G from 2G for the write tasks. For the read tasks, we will limit them to
3 minutes each. Finally, we will remove the verify option for the read jobs
as it is not needed.