exclude kernel, kernel-core, kernel-debug, kernel-rt and kernel-rt-debug from dnf/yum updates
if a test has kernel or kernel-core defined on metatada and there is a
newer version of kernel available in the repos restraint will install
the new kernel. If a test reboots the machine the machine will boot on
new kernel and not on cki kernel.

related: https://gitlab.com/cki-project/kernel-tests/-/issues/949