kunit: Fix the parsing for the new format adopted upstream
Some tests now produce subtest results. The parser was not configured
to work with this new format and reports as a failure.
Fix this by changing the removal of `lines starting with a whitespace`
with `only removing the first tab`. This allows the tappy tool to
properly parse the results.

Signed-off-by: Nico Pache <npache@redhat.com>