Add backup reviewers
For tests used by CKI, the current 'Maintainers' section in CODEOWNERS
will create approval rules that mostly contain only one person on them,
the maintainer. This might cause MRs to get blocked if that person is
not available.

Add Bruno and Linqing as backup 'management-override' users to all those
rules. Require at least 3 reviewers to skip paths where the maintainer
is not a member of redhat/centos-stream/tests/kernel yet.

Signed-off-by: Michael Hofmann <mhofmann@redhat.com>