ltp/include: protect systemd and restraint from oom killing
As we all decide to combine ltp wrapper into this one, the DEV team will shift
to use this one in smoke test as well.

But with no rpm package built in this repo, we suggest they go with
restraint way for submitting jobs. So protecting restraint and systemd
becomes very necessary.

Signed-off-by: Li Wang <liwang@redhat.com>