i2c-tools: sanity: Add tmt support for i2cdetect-smoke tests.
Although this test contains the `metadata` file which makes it functional
with restraint through beaker, tmt support is not present due to the lack
of the `main.fmf` file.  Add this to add support for the tmt.

Also, while at it, fix the permissions to 0755 from 0644, since tmt complains
for not having enough permissions at the entry point, that is `runtest.sh`.

Signed-off-by: Emmanuil Chatzipetru <echatzip@redhat.com>