Increase stalld test iteration runtime
The current runtime of 90s cuts the test very close in the cases where a
thread needs to get boosted twice to run. Sometimes this results in a
spurious error. Increase the stalld test time to make this unlikely to
happen. NOTE: It is possible for the test to take 3x cycles, but this
should be very unlikely.

Signed-off-by: Waylon Cude <wcude@redhat.com>