stalld typo fixup and max runtime fixup
Corrected a typo in a comment, changed the test max runtime to 30m (it
should take only 20m to complete, but give a 10m buffer for installing
dependencies), and added a bonus comment about how I know the test is
working. It would be possible to add a bonus check in the test related
to this as well, to check stalld's output to make sure the process is
boosted, but it seems unnecessary for now.

Signed-off-by: Waylon Cude <wcude@redhat.com>