Fixup on sched-psi memory test
This commit changes the memory pressure part of the sched-psi test. It
has two components. One is a C program that I wrote for the test, that
will create memory pressure by mmaping large chunks of memory. This
uses multiple threads to apply more pressure and to make it faster.

The other part, in the modified python file, checks for memory pressure
every 5 seconds and marks the test as passing if it sees any memory
pressure at all. This is necessary because memory pressure is ephemeral.
In all cases that I've found, once memory pressure is created the oom
killer starts killing processes, making the memory pressure go back to
0. This is one of the reasons the old test was failing so often, there
would be a spike in memory pressure early on in the runtime of
stress-ng, and then it would go away before the test measured the
pressure at the end of the 40 second run. The python script now will
check memory pressure over the course of 10 minutes, and will exit as
soon as memory pressure is detected and the test is guaranteed to pass.
This gives high memory systems plenty of time to mmap everything to
create pressure; but the test can finish early on low memory systems,
keeping the test time low.

Signed-off-by: Waylon Cude <wcude@redhat.com>