kpkginstall: use the right compiler for the cross compiling workaround
The build system sets various config options based on the
compiler/linker used to build the kernel. These config options can
affect the build process, so we need to ensure that the right
compiler/linker (i.e. gcc vs. clang) is used for the cross compiling
workaround.

Without this the workaroud may, for example, fail to rebuild a tool that
is actually needed to build a kernel module when the correct
compiler-linker combo is used, leading to test failures.

Should partially fix #1507

Signed-off-by: Ondrej Mosnacek <omosnace@redhat.com>