Use bash.orig for shellspec
When running shellspec tests in the CKI container images, two shells
might be available: the originally shipped bash version from rpm and a
backport for CKI. If the backport is present, it takes the place of
/bin/bash, and the original bash is available at /bin/bash.orig.

For the case where both shells are available, shellspec tests in
kernel-tests should use the original bash as the CKI bash is not avaible
on the VM/machine where the code under test will eventually run.

Towards that end, enable USE_BASH_ORIG that will automatically use
bash.orig if available.

Signed-off-by: Michael Hofmann <mhofmann@redhat.com>