fork of /distribution/kernelinstall Beaker task
The /distribution/kernelinstall task has a dependency on /kernel/include
and restraint cannot fetch-from-git dependencies from a different git
repo [*], so fork the task and store them both in the same kernel tests
repo.

Also copy the /kernel/include task files over as a dependency, but
rename the directory to kernel-include so restraint doesn't get
confused when extracting the repo tarball.

[*] https://github.com/restraint-harness/restraint/issues/140