distribution/kernelinstall: support kernel-64k-debug
Consider the runkernel var:
5.14.0-265.el9.aarch64+64k-debug -> 5.14.0-265.el964k-debug
5.14.0-265.el9.aarch64+64k -> 5.14.0-265.el964k
5.14.0-265.el9.aarch64 -> 5.14.0-265.el9

This change will make CheckKernel() handle those directly. We pass the
correct KVAR parameter directly to the function. For 64k debug, we pass
the KVAR with '64k-debug', for 64k non-debug, we pass it with '64k-up'.

With such handling, 5.14.0-265.el9 and 5.14.0-265.el9+64k could be
distinguished by the CheckKernel().

Signed-off-by: Chunyu Hu <chuhu@redhat.com>