general/include: remove kernel from requirement
when running tests with kernel-64k cki build, the --repo-post provides
the kernel-64k repo, there's no kernel packages avaialbe in the repo, as
a result. kernel-(modules,core) packages would be failed to download
defined in the metadata. It's not necessary to have kernel in
requirement at all, so remove it.

Signed-off-by: Chunyu Hu <chuhu@redhat.com>