kernelinstall: add MAX_REBOOT_TIME
The value for check of time kernel took to boot is hardcoded and there
is no way to override it.

This adds MAX_REBOOT_TIME parameter that specifies maximum time in
seconds allowed a kernel to boot.

This patch also removes duplicated check max boot time and removes
logging to debug log, since it is a duplication of the info.

Signed-off-by: Jan Tluka <jtluka@redhat.com>