xfs/xfstests: run old xfstests for old releases
Based on the kernel version we are running on, checkout
different branches of xfstests to test. Now we're running
the latest upstream xfstetss, this could cause false
alarms as new tests failing on old releases.
GFS2 is not included for now, because old versions may
not support GFS2 well enough.

Signed-off-by: Murphy Zhou <jencce.kernel@gmail.com>