libhugetlbfs: adjust the system clock immediately
It seems some of aarch64 platforms reboot on waiting too long for
chrony to synchronize system clock, so that rebuild package failed
with the Makefile wrong timestamp:

  make[1]: Warning: File 'Makefile' has modification time 16862 s in the future
  make[1]: Warning: File 'Makefile' has modification time 16862 s in the future

This patch introduces proactively correct the system clock to get
rid of the wrong time implications. If unable to correct, restart
the chronyd service completely.

Signed-off-by: Li Wang <liwang@redhat.com>