Ensure the Git package is installed
If Git has not already been installed by another test, such as when this
test is run independently, the setup phase reports the following error:

    runtest.sh: line 50: git: command not found

If the Git package is not installed (or the cloned directory is
otherwise unavailable on the filesystem after the clone operation
completes) then 'runtest.sh' will recursively call 'make' within the
current test directory instead of the cloned directory. The recursive
calls will continue until the system runs out of memory and the test hangs:

    CONFIG_MEMORY_FAILURE=y
    CONFIG_HWPOISON_INJECT=m

    ::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::
    ::   Setup
    ::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::

    Getting mce-test suit ...
    runtest.sh: line 50: git: command not found
    make[2329]: Entering directory '/mnt/tests/gitlab.com/redhat/centos-stream/tests/kernel/kernel-tests/-/archive/main/kernel-tests-main.tar.gz/general/memory/function/memory-failure'
    test -x runtest.sh || chmod a+x runtest.sh
    ( set +o posix; . /usr/bin/rhts_environment.sh; \
            . /usr/share/beakerlib/beakerlib.sh; \
            . runtest.sh )
    /usr/share/beakerlib/infrastructure.sh: fork: Cannot allocate memory
    cat: memory exhausted

Ensure that the dependency on Git is met by listing it explicitly.

Signed-off-by: Eddie Kovsky <ekovsky@redhat.com>