Add specific package names to metadata
The Restraint metadata file only allows us to specify package names
needed to run a test. It would be preferable to install the package
group 'Development Tools', but specifying a group name is not
supported.

In addition to the packages required for our main test, we also need to
ensure that the dependencies required by mce-test and mce-inject are
also installed.

Signed-off-by: Eddie Kovsky <ekovsky@redhat.com>