Verify cloned directory exists
If the 'git clone' operation fails (or the new directory is
otherwise unavailable on the filesystem) then the next 'make' commands
will execute in the current directory. This is not what the code intends
and can have severe side effects.

Ensure the directory change was actually successful, and report a
meaningful error message if it failed before aborting the test.

Signed-off-by: Eddie Kovsky <ekovsky@redhat.com>