Remove package 'kernel' as a test dependency
Specifying that the 'kernel' package should be installed before running
the test creates a conflict with the naming convention for kernel
variants.

When a test kernel is generated by the build system a unique pipeline ID
is appended to the package string. After the system has booted a
standard kernel 'dnf' can reinstall the kernel package. But if the system
has booted a kernel variant for testing 'dnf' has no way to resolve the
pipeline ID in the package string to install a standard kernel. This
results in the test aborting during the setup phase:

    ** Installing dependencies
    use_pty:FALSE rstrnt-package install kernel gcc

    Package kernel-5.14.0-284.11.1.el9_2.aarch64 is already installed.
    Package gcc-11.3.1-4.3.el9.aarch64 is already installed.
    Error:
    Problem: cannot install the best candidate for the job
    - nothing provides kernel-core-uname-r = 5.14.0-284.20.1.724_902851815.el9_2.aarch64 needed by kernel-5.14.0-284.20.1.724_902851815.el9_2.aarch64
    - nothing provides kernel-modules-core-uname-r = 5.14.0-284.20.1.724_902851815.el9_2.aarch64 needed by kernel-5.14.0-284.20.1.724_902851815.el9_2.aarch64
    - nothing provides kernel-modules-uname-r = 5.14.0-284.20.1.724_902851815.el9_2.aarch64 needed by kernel-5.14.0-284.20.1.724_902851815.el9_2.aarch64

The root cause of this issue has been identified in the build system:

    https://bugzilla.redhat.com/show_bug.cgi?id=2190323

However, specifying the 'kernel' package as a test dependency shouldn't
be necessary at all once the system has booted.

Signed-off-by: Eddie Kovsky <ekovsky@redhat.com>