Kdump: add test case for bz2211262
This issue was found on arm64 machine with more than 4 cpus that cpu#1 is offlined,cpu#2 is failed to get stopped and cpu#3 is panicked.
The expected results:
(1). Mapping between NT_PRSTATUS notes and cpu numbers is correct that CPU#1 and CPU#2 don't have NT_PRSTATUS notes because
CPU#1 is offlined and CPU#2 is failed to get stopped due to the irq mask.
(2). Crash utility never results in segmentation fault.

Signed-off-by: Xiaoying Yan <yiyan@redhat.com>