distribution/kernelinstall: abort task if SelectKernel fails
There is a case where the kernelinstall task would let any subsequent
tasks run even if the kernel is not properly installed.

When YumInstall and BrewInstall options are tried and would not succeed,
the SelectKernel is checked and it returns with error. This error is
reported back to Beaker, however since no reboot is called, the task
continues with no further CheckKernel check. This check is typically called
after reboot and task is aborted if the check fails.

To fix this, we should call rstrnt-abort immediately when SelectKernel fails.

Signed-off-by: Jan Tluka <jtluka@redhat.com>