Remove use of logdir from rteval unit test
This test uses an rteval command line flag that no longer exists, or
maybe never existed. The -L flag, that turns on logging, does not take a
logdir like the test assumes. This was causing the test to fail.

Remove use of logdir, it doesn't look like it's used anywhere.

Signed-off-by: Waylon Cude <wcude@redhat.com>