Add guest poweroff test for x86_64
This test checks for regressions in poweroff behavior on x86_64. It is
not a "utility" test to execute a shutdown as part of a larger recipe.
The host issues the poweroff command to a virtual guest. This way the
host can still monitor the state of the guest system after it powers
off.