kvm-selftests build from upstream repo
This commit will fix the issues with kvm-selftests during the kvm
upstream testing. Before this change, the script presumed that all
upstream testing has pre-builded kernel-selftests and the
`CKI_SELFTESTS_URL` variable is available. This caused issues when the
`CKI_SELFTESTS_URL` wasn't set. This change use `K_IsKernelRPM` function
from kernel-include to detect how the kernel was installed and if it is
needed to build the selftests or not.

Signed-off-by: Jan Richter <jarichte@redhat.com>