[rt] rteval: do not set load or measure cpu flags if isolated cores is empty
There is a possibility that the user requests running either rteval load or measurement threads on isolated cores, but the system was not configured with any isolated cores.

The previous behavior would set these flags with no cpulist, which would be in error.  Instead, simply log that these flags cannot be set to isolated cores when none are set in the system.

Signed-off-by: Mike Stowell <mstowell@redhat.com>