[rt] rteval: unset FLAG_LOADS and FLAG_MEASURE if either is empty
In the event the user requests that either measurement or loads cpulist utilize isolated cores but none are available on the SUT, unset both parameters.

This is necessary for two reasons:

(1) A parameter expecting a value cannot be passed to a command without a value set, and
(2) In the likely event one flag was requested with isolated cores and the other with housekeeping cores, the parameter that requested housekeeping cores would monopolize all CPUs on the system and lead to an rteval IndexError since the other parameter has no other cores to utilize.

For example, on a system with 16 cores, none isolated:

~$ rteval --duration=10m -D -L --loads-cpulist 0-15
** ERROR **  Failed to save log entry
SMBIOS3: entry point at 0x698a6000
got system topology: 2 node system (8 cores per node)
Traceback (most recent call last):
  File "/usr/bin/rteval", line 346, in <module>
    msrcfg.cpulist = compress_cpulist(tmplist)
  File "/usr/lib/python3.9/site-packages/rteval/systopology.py", line 164, in compress_cpulist
    if isinstance(cpulist[0], int):
IndexError: list index out of range

Signed-off-by: Mike Stowell <mstowell@redhat.com>