rt-tests/rteval/runtest.sh : don't mount debugfs twice
On some platforms, trying to mount debugfs twice will fail.  Since
this is using the device "none" this leads to the red-herring error
something like

  [root@localhost ~]# mount -t debugfs none /sys/kernel/debug/
  mount: /sys/kernel/debug: none already mounted on /sys/kernel/tracing.

This is actually because debugfs is already mounted; but it's getting
confused and telling you about /sys/kernel/tracing (or bpf, or
whatever else might be mounted on "none").

Closes: https://issues.redhat.com/browse/RHEL-8856