memfd_create: add process state wait
The purpose of this change is to address a potential race condition
where the test proceeds to check for seals on the file descriptor
before the background process has applied them.

By adding the `t_process_state_wait $! S` command, the script ensures
that the t_memfd_create process is in the appropriate state, which
should minimize the timing issues and make the test results more reliable,
especially on systems like s390x where syscall delays due to overload
have been identified.

Signed-off-by: Li Wang <liwang@redhat.com>