perf - internal-testsuite: Waive shadow and deamon subtests
For subtest shadow stat there was a very rare rounding error
discovered in rhel9.3 which was fixed in rhel9.4. The fix wasn't
backported to older versions and the test may be causing failures.

The daemon subtest is unstable for s390x machines where there was
minimum cases of this test failing. However, the fail wasn't possible to
reproduce manually. As there is not enough information to try to resolve
why this fail is happening, waive the test for s390x as an unstable
test.