memory/kmemleak: relocate the status file to /mnt
kmemleak test did more reboot then expected. This is because when using
rpm for testing, the test case is not cleanuped if the test installed
more than once, but in gitlab case, everytime a fetch of the test case
will result to the cleanup of the old test files. kmemleak test will
detect leak after every normal beaker task, but it will do reboot every
time it's wasty.

Signed-off-by: Chunyu Hu <chuhu@redhat.com>