perf - internal-testsuite: Fix regex for bz tests
The allowlisting of these tests(bz1414043 coverage, bz1308907 coverage)
was not working correctly. The regex wasn't removing the numbering as
there remained an unwanted space in front of the test number as those
are aligned on the colon. Fix the regex to correctly match the test
description and allow waiving the tests.