kernelinstall: check curl downloads
In the linked issue, we diagnosed a problem with a test building
modules that actually stemmed from the initial installation steps
here.  It seems that a download of the -devel package for the kernel
under test failed from brew, and it was not installed.  Tests that
later on relied on "kernel-devel" had that satisfied by the
base-kernel package, and they failed trying to build modules against
the kernel under test.

There are many points that probably should fail the test; however to
keep things tractable this introduces a checked download function and
replaces any suitable unchecked curl calls with it.  This should
ensure we at least have the files on disk.

We add some retry options just to give it a chance to recover from
basic transient errors; but the important thing is that it should
error out clearly on failure.  We caputre the output to the debug log
file so we can debug failures too.

Closes: https://issues.redhat.com/browse/RHEL-27119