distribution/kernelinstall: fix for recent upstream kernel changes
The rpms built with `make srcrpm-pkg` after version 6.8.0 no longer contain
the vmlinuz along with the version in the file list. The upstream kernel moved
copy of the VR file to /boot in %post. Because of this, the candidate list
of vmlinuz files is empty and the task fails to verify the kernel is installed.

The fix is to check the different location.

Signed-off-by: Jan Tluka <jtluka@redhat.com>