LTP: upload dmesg log of failed test cases
dmesg log is only being uploaded if a failure is detected on dmesg. In some
cases a testcase can fail and no dmesg failure is reported, still having the
dmesg log in this case can be helpful.

Note: restraint doesn't support multiple logs to be uploaded to a subtest,
therefore upload it as a global file.