.gitlab-ci.yml : don't suggest .zip archive
If you use the .zip archive, you don't seem to get correct file
permissions in the extract files of your beaker job.

I think what is happening here is that <fetch> is streaming the result
via curl to libarchive [1] which, since .zip is not a streaming format
and keeps the permissions info at the end of the file, results in
permissions essentially being ignored [2].  As is usually the way,
once I had debugged all this and knew what to look for, I found that
restraint is aware of the problem too [3] :)

I can quite easily replicate this with two jobs, one using a .zip that
fails with permissions errors trying to start 'runtest.sh' and another
using .tar.gz that doesn't.  I think this may be mostly responsible
for

 $ git grep 'chmod a+x ./runtest.sh' | wc -l
 64

in this source repository.

I imagine this was both copied from somewhere else, and in turn has
been copied far and wide.  I think changing the suggestion to .tar.gz
will at least help here.  As a bonus, it seems to be half the size

 ... 41M Apr 17 07:10 archive.tar.gz
 ... 90M Apr 17 08:41 archive.zip

[1] https://github.com/restraint-harness/restraint/blob/master/src/fetch_uri.c
[2] https://github.com/libarchive/libarchive/issues/1106\
[3] https://github.com/restraint-harness/restraint/issues/136