process/reg-suit: 1157802: report skip when nr_cpu < 4
When skipping the test, we report the result as 'SKIP'. Add the
'STATUS_FILE' variable for sub functions to pass the sub func result to
the main loop. If the STATUS_FILE content is SKIP, the main loop report the test
as skip. Handling in the main loop will avoid duplicated results
for same sub test, (such as 'bz1157802 PASS' and 'bz1157802 SKIP both
appear)

Signed-off-by: Chunyu Hu <chuhu@redhat.com>