mmra/resident_memory: read memory info from smaps instead of status
read Rss and Locked from /proc/$pid/smaps instead of
/proc/$pid/status.

status file can generate false positive as it is not guaranteed
to be updated everytime the file is opened.
smaps file gives this guarantee but now statistics must be found
for a specific memory mapping in the process.

Add more detailed logs in case of test success/failure.