kunit: generate test list by scanning modules
Drop the kunit-tests.list file for a better approach.
We can scan the lib/modules folder for modules that contain kunit
symbols. We used to have to keep a file up to date with all the kunit modules
names. Instead, we will now scan the modules internal folder, and look
for symbols related to kunit. This not only removes some
maintainership burden, but also automatically includes new tests in
the ark suite, which will give us faster results for new tests.