kvm-unit tests early logging
This commit moves the uploading of the test logs just after the test
run. This is important in case that the kvm-unit test-suite will
timeout. With the current solution, when the script is interrupted due
to the timeout the logs are not uploaded and the DW is not able to
evaluate known bugs. Uploading the logs right after the test run will
fix this issue.

Signed-off-by: Jan Richter <jarichte@redhat.com>