kvm-selftests RHEL10 compilation fix
Some kvm-selftests are failing on RHEL10 due to usage of incompatible
gcc flags during the test build. The fix for this has been sent to
upstream,  but until it will be rebased to RHEL10 we need a workaround
to make those tests running.

Reference: https://issues.redhat.com/browse/RHEL-58930
Signed-off-by: Jan Richter <jarichte@redhat.com>