leap_hrtimer: Make thread callback CPU unaware
Make it so the threads have no idea about what CPU they're on,
and leave it to the main thread to handle that.

This way each thread only concerns itself with checking if it slept
properly, and let's the main thread figure out which thread (and
therefore CPU) failed. With this approach the thread's return value
can be checked directly instead of having a global array of failures
(aka "p" before), and the thread doesn't need to be passed any argument.

Signed-off-by: Andrew Halaney <ahalaney@redhat.com>