rt_ssdd: fix the result report logic and simplify the test
log_pass / log_fail don't report the results to restraint, all they do is adding
a LOG line in the default output file (stdout). And the command enclosed in
`run` is trailed with a `tee` operation, which always returns true.

Beside of fixing the result problem, since ssdd has a problem of not able to
exit itself, in order to not let it abort the recipe or killed by watchdog, this
commit also adds a `timeout` for each ssdd command so that when ssdd can't exit
we can intentionally kill it from the test.