vm/ltp-hugetlbfs: set max_time to 90m
if there's no 'max_time' in metadata, it would be timeout as 30m by
default, it's good for most cases, but if run into a kvm guest with 1
processor, it would be external watchdog expired and test aborted.

provide it a 90m to make it be able to finish on 1 processor machines.

Signed-off-by: Chunyu Hu <chuhu@redhat.com>