internal-testsuite/runtest.sh: Fix checking the result
The result line was changed in perf and the current regex wasn't able to
correctly parse the result meaning that the tests were always passing
through the else branch. Switch the if and else branches so that if
there would be any changes in the result printing in perf and the
wrapper wasn't able to parse the result, the tests will rather all fail
then pass.

Signed-off-by: Veronika Molnarova <vmolnaro@redhat.com>