internal_testsuite/runtest.sh: Fix result checking for subtests
The result printing was changed for subtests as well when it doesn't
contain the name of the main test. The wrapper was also incorrectly
interpreting possible regex characters in the test name, which caused
issues. Now search for any of the possible test or subtest name in the
output while not using anyy regex characters.

Signed-off-by: Veronika Molnarova <vmolnaro@redhat.com>