Fixed kernel version guessing
The previous method was not specifying a specific package name and
when invoked in machines with more than one kernel-core package
installed the output was the concatenation of all installed versions.

Now we get current kernel running using uname so we can be more
specific in the kernel-core package name and get only the given version.