VMMTIER1_old config file for LTP generic tests
The VMMTIER1_old test in 7.9.z fails because it cannot locate the VMMTIER1_old config file.
This file was removed when LTP was moved to the public kernel tests repository, and recently,
we started using the public version of LTP/generic tests in Z-stream testing.

This commit restores the missing VMMTIER1_old config file by copying it to the new test repository,
ensuring that the test can locate the required configuration and execute correctly.

Fixes: FileNotFoundError for VMMTIER1_old.log and related failures in Z-stream testing.