memory-failure: rearrange the test with beakerlib and skip unsupported arches properly
Resolves: RHELTEST-87

mcelog is only applicable on x86_64 and i386, and the test basically can only be
run on x86_64 and ppc64le, skip the unsupported scenarios properly.

Additionally, this commit uses beakerlib functions to handle the assertions
correctly in a relatively clearer fashion, verbosity is enhanced as well to
facilitate debugging.

Signed-off-by: Shizhao Chen <shichen@redhat.com>