security/selinux/testsuite: use newer version of test suite to pass on rhel-10
https://gitlab.com/redhat/centos-stream/tests/kernel/selinux-testsuite/-/commit/6aba9eb5ba78bcabe8abb5e3809f3680dc0403a1

In CentOS Stream 10 (and possibly other environments) the
net.mptcp.enabled sysctl will be set to 0 by default. The testuite
currently doesn't detect this and hangs when trying to run the MPTCP
tests. Adjust the test to temporarily enable MPTCP during the test so
that it can succeed in these conditions.