Changes test file for qatzip test
- The previous approach used randomized data for compression testing of
qatzip. However, the compressed file results in a larger file size
than the original file. This is a known issue with random data
compression, thus is not specific to qatzip.

- Decompression does not get corrupted even though random data gets
inflated after compression. Regardless, this commit changes the test file
for qatzip simply to adhere to the usecase of compression tools such as
qatzip.

- The new approach uses the same file which is used for the qat-zstd-plugin
test found at https://github.com/yewq/Silesia-compression-corpus.

Signed-off-by: Denis Aleksandrov <daleksan@redhat.com>