kselftest: only run test concurrently when WORKERS is supplied
Commit 4a1e61febe27 ("Add CONCURRENT_WORKERS variable to RunKSelfTest to
allow concurrent testing") enabled all test cases to run concurrently when
the WORKERS variable is set. However, it no longer properly checks the test
case return value. The ret is always 0, as the previous operation was a cat,
which falsely reports all tests as PASS.

To address this, we need to track all the pid return values. At the same
time, I would prefer to run the test separately when WORKERS is not
supplied, as I may use run_kselftest.sh to run the selftests in future.

Signed-off-by: Hangbin Liu <haliu@redhat.com>