distribution/kpkginstall: fix matching of KVER in repoquery output
Currently the test doesn't expect boolean expressions in repoquery output.
Since we recently introduced a modules-extra-matched, it breaks the parsing
as it matches the first line from this output:
  # rpm -qlp --provides --requires kernel-64k-6.15.0-0.rc3.34.3745_1789572255.eln.aarch64.rpm | grep uname-r
  ((kernel-64k-modules-extra-uname-r = 6.15.0-0.rc3.34.3745_1789572255.eln.aarch64+64k) if kernel-modules-extra-matched)
  kernel-64k-core-uname-r = 6.15.0-0.rc3.34.3745_1789572255.eln.aarch64+64k
  kernel-64k-modules-core-uname-r = 6.15.0-0.rc3.34.3745_1789572255.eln.aarch64+64k
  kernel-64k-modules-uname-r = 6.15.0-0.rc3.34.3745_1789572255.eln.aarch64+64k

and KVER ends up as "6.15.0-0.rc3.34.3745_1789572255.eln.aarch64+64k) if kernel-modules-extra-matched)"

Change the pattern, so it doesn't match boolean expressions, and also just look
at core-uname-r line.

Signed-off-by: Jan Stancek <jstancek@redhat.com>