[rt] add a regression test for printk
This is a regression test for RHEL-72528, it serves effectively as a test to
stress the printk subsystem, it does so by creating a large number of memory
cgroups and launching many processes that intentionally exhaust their cgroup's
memory limit, triggering the OOM killer. At the same time, the test continuously
writes long messages to /dev/console to generate heavy printk activity.

The test passes if the system remains responsive and no deadlock occurs during
the stress period.

Signed-off-by: Shizhao Chen <shichen@redhat.com>