[rt] stalld: make the test deterministic by using event-based checks
The previous implementation relied on a timing-based threshold. It measures the
total execution time of the NORMAL task and passes if it completes within an
"arbitrary" time limit. The approach is flaky as it's hard to predict the
accurate time it takes for the starved task to complete. Timing varies
significantly across different platforms, it's also affected by the system
status at that period. No guanrantee can be made that a task can "complete"
within a known number of boosts.

However, we can be sure that a task is boosted at a specified time frame. This
commit refactors the test with ftrace by checking whether there were any
sched_switch events of the starver task during the starvation period. The stalld
arguments are made very generous to make sure that the boost can happen
frequently and quickly after the starvation happens.

Additionally, a workaround is added to prevent the critical ktimers thread from
starvation with the new log-only mode stalld and fair-server.

Resolves RHELTEST-857

Signed-off-by: Shizhao Chen <shichen@redhat.com>