Verify a 4k kernel is not installed alongside the 64k kernel
JIRA: https://issues.redhat.com/browse/RHEL-82309

Installing the operating system with a 64k kernel should not install the 4k
kernel. It is not recommended to move between 4k and 64k kernels without
reinstallation of the operating system. Additional packages should not bring
in the 4k kernel post installation.

Test that the state of the operating system has not changed after
installing with a 64k kernel.

Here is an automated test run through Beaker reporting a test failure
because the 4k and 64k kernel are both installed on the system:

    https://beaker.engineering.redhat.com/jobs/11265297

This test has also been manually verified on several systems using
different page sizes:

orin with 64k and 4k kernels
---
    [root@orin ~]# ls /boot/vmlinuz*
    /boot/vmlinuz-5.14.0-590.el9.aarch64  /boot/vmlinuz-5.14.0-592.el9.aarch64+64k

    [root@orin ~]# ./pagesize
    Architecture is: aarch64
    Operating system release: 5.14.0-592.el9.aarch64+64k
    Page size is: 65536

    Fail: a non-64k kernel was found on /boot: vmlinuz-5.14.0-590.el9.aarch64

x86_64 Fedora with a 4k kernel
---
    ❯ ls /boot/vmlinuz*
    /boot/vmlinuz-0-rescue-5c9eaa98f620496ab28dfc031f926829*  /boot/vmlinuz-6.14.6-200.fc41.x86_64*
    /boot/vmlinuz-6.14.5-200.fc41.x86_64*                     /boot/vmlinuz-6.14.9-200.fc41.x86_64*

    ❯ ./pagesize
    Architecture is: x86_64
    Operating system release: 6.14.9-200.fc41.x86_64
    Page size is: 4096

    Fail: This test should only be run on aarch64 with a 64k kernel

grace hopper with only a 64k kernel
---
    [root@nvidia-grace-hopper-09 ~]# ls /boot/vmlinuz*
    /boot/vmlinuz-0-rescue-c076eaf714194170b6cf47e93c653a1d  /boot/vmlinuz-5.14.0-593.6998_1867700032.el9.aarch64+64k

    [root@nvidia-grace-hopper-09 ~]# ./pagesize
    Architecture is: aarch64
    Operating system release: 5.14.0-593.6998_1867700032.el9.aarch64+64k
    Page size is: 65536

    [root@nvidia-grace-hopper-09 ~]# echo $?
    0

ampere with only a 4k kernel
---
    [root@ampere-mtsnow-altramax-02-vm-14 ~]# ls /boot/vmlinuz*
    /boot/vmlinuz-0-rescue-9faf9db8aefa4cf799122f82ea9b67e9  /boot/vmlinuz-6.12.0-95.el10.aarch64

    [root@ampere-mtsnow-altramax-02-vm-14 ~]# ./pagesize
    Architecture is: aarch64
    Operating system release: 6.12.0-95.el10.aarch64
    Page size is: 4096

    Fail: This test should only be run on aarch64 with a 64k kernel

Signed-off-by: Eddie Kovsky <ekovsky@redhat.com>