scheduler/sched_nohz_tick: undefine sched_attr when defined by glibc-headers
sched_attr, sched_setattr() and sched_getattr() are defined by
glibc-headers since glibc-2.34-176.el9 9.7. it will cause compile failure
caused by redefinition errors. So remove the double definition in the
app when it's defined in the glibc-headers to fix the compile issue.

This is a workaround, need to keep an eye on cyclictest updates or glibc
udpates int the next.

Signed-off-by: Chunyu Hu <chuhu@redhat.com>