internal-testsuite: Strip '(exclusive)' from test names
The word '(exclusive)' is used as an identification of tests that should
_not_ be run in parallel in case the testsuite is run in parallel. The
upstream developers decided to put it into the testcase's name at the
end, however, that, of course breaks our mechanism for ignoring known
failures, since it is based on the name (as the test number is not
reliable).

Strip the '(exclusive)' "suffix" from the test names if present to fix
this problem.

Signed-off-by: Qiao Zhao <qzhao@redhat.com>